backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.15.4
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.8
importlib_metadata==8.4.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
joblib==1.4.2
keras==2.12.0
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
# opencv-python==4.9.0.80
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
owlready2==0.46
packaging==24.1
pandas==1.5.3
pillow==10.3.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
segment-anything==1.0
six==1.16.0
sklearn==0.0
sympy==1.13.2
tbb==2021.13.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.20.1
